+++ cfglp0-src-int/level-2-64bits/local-environment.hh
@@ -49,8 +49,8 @@ protected:
 	Result_Enum result_type;
 
 public:
-	virtual float get_value();
-	virtual void set_value(float value);
+	virtual double get_value();
+	virtual void set_value(double value);
 
 	virtual bool is_variable_defined();
 	virtual void set_variable_status(bool def);
@@ -62,8 +62,8 @@ public:
 class Eval_Result_Value:public Eval_Result
 {
 public:
-	virtual void set_value(float number) = 0;
-	virtual float get_value() = 0;
+	virtual void set_value(double number) = 0;
+	virtual double get_value() = 0;
 
 	virtual bool is_variable_defined() = 0;
 	virtual void set_variable_status(bool def) = 0;
@@ -80,8 +80,8 @@ public:
 	Eval_Result_Value_Int();
 	~Eval_Result_Value_Int();
 
-	void set_value(float number);
-	float get_value();
+	void set_value(double number);
+	double get_value();
 
 	void set_variable_status(bool def);
 	bool is_variable_defined();
@@ -92,14 +92,14 @@ public:
 
 class Eval_Result_Value_Float:public Eval_Result_Value
 {
-	float value;
+	double value;
 	bool defined;
 public:
 	Eval_Result_Value_Float();
 	~Eval_Result_Value_Float();
 
-	void set_value(float number);
-	float get_value();
+	void set_value(double number);
+	double get_value();
 
 	void set_variable_status(bool def);
 	bool is_variable_defined();
@@ -116,8 +116,8 @@ public:
 	Eval_Result_Value_Double();
 	~Eval_Result_Value_Double();
 
-	void set_value(float number);
-	float get_value();
+	void set_value(double number);
+	double get_value();
 
 	void set_variable_status(bool def);
 	bool is_variable_defined();
